Prepare for release

+++ apk/CONTROL/start-stop.sh
@@ -7,13 +7,17 @@ cd ${APKG_PKG_DIR:-/nonexistent} || exit 1
 
 case $1 in
   start)
-    logger "[Persistence] Activating user configuration..."
+    logger "[${WHAT}] Activating user configuration..."
     touch "${APKG_PKG_DIR}/active"
     ./CONTROL/start-hook.sh
     ;;
 
   stop)
-    logger "[Persistence] Removing user configuration..."
+    # Don't remove to re-install the files, with double docker restart
+    if test "x${APKG_PKG_STATUS}" == "xupgrade"; then
+      exit 0
+    fi
+    logger "[${WHAT}] Removing user configuration..."
     rm -f "${APKG_PKG_DIR}/active"
     ./CONTROL/stop-hook.sh
     ;;
@@ -28,9 +32,9 @@ case $1 in
     # Do not switch files off/on in sequence, just do it once.
     # Otherwise it'll restart docker every time
     if test -f "${APKG_PKG_DIR}/active"; then
-      ./CONTROL/start-hook.sh
+      ./CONTROL/start-stop.sh start
     else
-      logger "[Persistence] Service is stopped, cannot reload."
+      logger "[${WHAT}] Service is stopped, cannot reload."
     fi
     ;;
